#115 refactor: 게시글 등록, 이미지 업로드 컴포넌트 스타일 분리
