+++ src/components/imageUpload/ImageUpload.jsx
@@ -1,6 +1,6 @@
 import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
-import { Button, Spin } from 'antd';
+import { style } from './imageUpload.style';
+import { Spin } from 'antd';
 import { PlusOutlined, UploadOutlined } from '@ant-design/icons';
 import { imageUploader } from '../../service/imageUploder';
 
@@ -101,46 +101,12 @@ const ImageUpload = ({ setImageUpload, pageName }) => {
 
 export default ImageUpload;
 
-const Wrap = styled.div`
-  margin-right: 20px;
-`;
-
-const ReviewInput = styled.input`
-  display: none;
-`;
-
-const PostWrap = styled.div`
-  width: 100%;
-  margin-top: 20px;
-  margin-bottom: 10px;
-`;
-
-const PostUpload = styled.div`
-  width: 100px;
-  height: 100px;
-  border: 1px dashed #bdbdbd;
-  transition: all 200ms ease;
-  cursor: pointer;
-
-  &:hover {
-    border: 1px dashed #73d13d;
-    color: #73d13d;
-  }
-`;
-
-const UploadContent = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  height: 100%;
-  padding: 20px;
-`;
-
-const PostInput = styled.input`
-  display: none;
-`;
-
-const ButtonContent = styled(Button)`
-  width: 148px;
-`;
+const {
+  Wrap,
+  ReviewInput,
+  PostWrap,
+  PostUpload,
+  UploadContent,
+  PostInput,
+  ButtonContent,
+} = style;
